msm: spm-devices: Set up Q2S configuration with SPM modes
ARM V8 targets that support Q-Channel interface can use either Q-Channel
interface to detect wake up requests and/or use the SPM to wake up the
core. Certain low power modes are more efficient when using just the
Q-Channel interface for wake up and some with SPM. Toggle the use of
Q-Channel and SPM interfaces using the Q2S configuration for the core.
The register for the configuration is a reg field of the SPM devices.

Signed-off-by: Praveen Chidambaram <pchidamb@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>