arm64: Fix leak issue when free DMA memory
In __free_from_pool() function the count calculation does not use
alligned page size so that the calculated page count is incorrect.
This results that the page bit will not be set to zero when free
that page which will cause the bitmap leak and makes DMA allocation
running out of bitmap.

Signed-off-by: Yue Ma <yuem@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>