msm: pm: Fix hotplug warmboot address for CPU0
On earlier targets, CPU0 is never hotplugged out and is always the last
core to suspend. So, the warmboot address of CPU0 was always set to
follow the idle path instead of resuming from secondary_startup.

On arch64 kernel, this assumption is no longer valid and CPU 0 could be
hotplugged out. In such a case, CPU0 might never be brought of hotplug.
Additionally during suspend, if Core 0  is hotplugged out, the warmboot
address of last core entering is incorrectly configured for
secondary_startup.

Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>