ARM: dts: apq8084: Add virt_ppi irq to bypass list
Add ARM arch_timer virt_ppi irq to the bypass list.

PPI irq is a per CPU irq, but the list in mpm_of is not per CPU variable.
It will be added to the prevent list when any CPU is up, and it will be
removed from the list when any CPU is down. So if the last cpu hotplug
operation is UP, then it will be added to the prvent list, and prevent
vddmin in idle path.

Signed-off-by: Ke Lv <a2435c@motorola.com>
Signed-off-by: Lianwei Wang <a22439@motorola.com>
Reviewed-by: Lian-Wei Wang <lian-wei.wang@motorola.com>
Reviewed-by: Christopher Fries <cfries@motorola.com>