cpuidle: lpm-levels: Flush L2 cache when hotplug happens before probe
If CPU hotplug is initiated before probe, the L2 of the cluster is not
configured for any low power mode and is not flushed. This results in data
incoherencies in ramdumps in particular, if the system crashed after.

To ensure that L2 is flushed, flush L2 during hotplug until lpm-levels
module is probed

Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>