msm: spm: Configure q2s based on qchannel property when entering PC
On msm8994v1, qchannel should not be ignored when entering power
collapse, due to hardware limitations. Add support for a boolean
property in the spm driver to identify these targets. If this property
is set, the spm driver will configure qchannel to be used when in power
collapse. On targets where this property is not specified, qchannel
will be configured to be ignored when in power collapse.

Signed-off-by: Karthik Parsha <kparsha@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>