ANDROID: dm: Add android verity target
This device-mapper target is virtually a VERITY target. This
target is setup by reading the metadata contents piggybacked
to the actual data blocks in the block device. The signature
of the metadata contents are verified against the key included
in the system keyring. Upon success, the underlying verity
target is setup.

BUG: 27175947

Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>